Added support for mod operation with integers and doubles and all combinations (see notes below for implementation details)
Needed to change the handling of mod, previously sympy was converting it to Mod and thence to mod. I found a workaround to get a single function to work for both integers and floats in C++ but not in Cython. So I changed it so that mod is left as % and now CPPNodeRenderer converts this to _brian_mod, which is always put into the support code (I created a new attribute universal_support_code to put things that are always in the support code).