Mark TimedArray function in C++ as "static inline"
According to the C99 standard that is used by default by clang on OS X,
functions can not only have an inline version, they need a non-inlined
version as well (http://clang.llvm.org/compatibility.html#inline).
This can lead to strange issues, e.g. an older implementation of the
function with the same name but referring to different variables can be
picked up, leading to the use of incorrect values from memory in the
TimedArray.