Fix an issue with indexing in `PopulationRateMonitor`
In C++-based targets, the stored length of the recorded timepoints/rates
were not correctly updated. Indexing with slices would then return empty
arrays. The fix is the same as we did a while ago for `SpikeMonitor`:
update "N" (the total number of time points/rates stored) from within
the template code.