Change test_long_run to not use  internals
The tests assumes a definition of a network variable in generated code,
which is not the case e.g. in brian2genn. The new test instead uses only
"standard" Brian code. It also does not run on brian2genn due to the lack
of multiple run support, but it correctly fails with a NotImplementedError
instead of during compilation.