Handle .ndim and [:, np.newaxis] on VariableView
These are used when handing a VariableView directly to matplotlib, e.g.
for spike_mon.i (an expression like spike_mon.t/ms directly results in
an ndarray and is therefore not concerned). This problem was introduced
with matplotlib 3.5.2, previous versions converted VariableView along
the way with a call to np.asanyarray.

Fixes #1409