Adapt to numpy's new copy keyword semantic
In numpy 1.x, `array(x, copy=False)` meant : do not copy if possible.
In numpy 2.0, it will raise an error if copying is not possible. This
commit replaces the use of `array(x, copy=False)` by `asarray(x)`, which
has the "do not copy if possible" in both numpy 1.x and numpy 2.0.