Avoid `ReferenceError` when using `run_args`
When using `run_args`, we go through all variables in the array cache, and check whether they should be invalidated. This can lead to a `ReferenceError` if a variable in the cache refers to an ephemeral subgroup. The fix is simply to swallow this error (but maybe these variables shouldn't be in the array cache in the first place?)

Fixes #1673